requirements_macos_mps.txt
# --- Core scientific stack (versions aligned with your YAML) ---
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
requests==2.32.3
pyyaml==6.0.1
networkx==3.2.1
sympy==1.12
typing_extensions==4.11.0

# --- From the pip: section in environment.yml ---
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
biopython==1.84
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
fair-esm==2.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
gudhi==3.10.1
importlib-resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.6
multidict==6.0.5
packaging==24.1
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
pyparsing==3.1.2
scikit-learn==1.5.1
scipy==1.13.1
sentry-sdk==2.11.0
setproctitle==1.3.3
smmap==5.0.1
threadpoolctl==3.5.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
yarl==1.9.4
zipp==3.19.2
